requirements.txt
agate==1.6.1
agate-dbf==0.2.0
agate-excel==0.2.2
agate-sql==0.5.3
asn1crypto==0.22.0
Authlib==0.10
awscli==1.15.82
Babel==2.5.3
beautifulsoup4==4.6.0
boto3==1.7.81
botocore==1.10.81
bs4==0.0.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.8.0
colorama==0.3.7
cryptography==1.9
csvkit==1.0.3
cycler==0.10.0
dask==1.1.4
dbfread==2.0.7
decorator==4.0.11
docopt==0.6.2
docutils==0.14
dryscrape==1.0
EasyProcess==0.2.3
editdistance==0.4
et-xmlfile==1.0.1
ez-setup==0.9
facebook-sdk==2.0.0
flake8==3.4.1
Flask==0.12.2
Flask-Cors==3.0.7
future==0.16.0
fuzzywuzzy==0.15.0
gspread==0.6.2
httplib2==0.10.3
idna==2.5
isodate==0.6.0
itsdangerous==0.24
jdcal==1.3
Jinja2==2.10.1
jmespath==0.9.3
kiwisolver==1.0.1
leather==0.3.3
lxml==3.8.0
MarkupSafe==1.0
matplotlib==3.0.3
mccabe==0.6.1
nameparser==0.5.2
networkx==1.11
numpy==1.13.1
oauth2client==4.1.1
oauthlib==2.0.2
openpyxl==2.5.2
pandas==0.20.3
pandas-access==0.0.1
parsedatetime==2.4
pdfminer3k==1.3.1
pew==1.1.0
Pillow==5.4.1
pipenv==8.2.7
pipreqs==0.4.9
ply==3.10
py==1.4.34
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycodestyle==2.3.1
pycparser==2.17
pyflakes==1.5.0
PyJWT==1.5.2
pyOpenSSL==17.0.0
pyparsing==2.3.1
PySocks==1.6.7
pytest==3.1.2
python-craigslist==1.0.6
python-dateutil==2.6.1
python-dotenv==0.10.0
python-Levenshtein==0.12.0
python-slugify==1.2.5
pytimeparse==1.1.7
pytz==2017.2
PyVirtualDisplay==0.2.1
PyWavelets==1.0.2
PyYAML==3.12
requests==2.21.0
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.13
scikit-image==0.14.2
scipy==1.2.1
selenium==3.12.0
six==1.10.0
slackclient==1.2.1
SQLAlchemy==1.2.6
style==1.1.0
tika==1.14.1
toolz==0.9.0
tweepy==3.5.0
twilio==6.4.1
Unidecode==1.0.22
update==0.0.1
urllib3==1.24.2
virtualenv==15.1.0
virtualenv-clone==0.2.6
webkit-server==1.0
websocket-client==0.47.0
Werkzeug==0.12.2
xlrd==1.1.0
xvfbwrapper==0.2.9
yarg==0.1.9
